+++ uber/src/main/java/com/nwtkts/uber/dto/HistoryRideDetailsDTO.java
@@ -23,7 +23,7 @@ public class HistoryRideDetailsDTO {
     @JsonFormat(pattern="yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime startTime;
 
-    private Integer calculatedDuration;
+    private Double calculatedDuration;
     private String routeJSON;
     private Double price;
