+++ uber/src/main/java/com/nwtkts/uber/dto/HistoryRideDetailsForDriverDTO.java
@@ -25,7 +25,7 @@ public class HistoryRideDetailsForDriverDTO {
     @JsonFormat(pattern="yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime startTime;
 
-    private Integer calculatedDuration;
+    private Double calculatedDuration;
     private String routeJSON;
     private Double price;
